requirements.txt
scrapegraphai>=1.40.1
scrapegraph-py>=1.12.0
python-dotenv>=1.0.1